0,2,2
Added !sus and !logs
rearranged every command
