+++ config.json
@@ -1,12 +1,11 @@
 {
-  "prefix": "!",
-  "token": "",
+  "token": "Nzk5NzUyODQ5MTYzNTUwNzIx.YAIJ4Q.wkqKeG6hELV5mjuhGX7Ht8n1r_w",
   "faliedCommandTO": 4000,
   "failedEmbedTO": 15000,
   "deleteFailedMessaged": true,
   "bot_info": {
     "name": "crazybot",
-    "version": "0.1.2",
+    "version": "0.2.0",
     "auhtor": "crazy4k",
     "authorID": "573205398526361620"
   }
