+++ functions/checkRoles.js
@@ -1,22 +1,27 @@
-module.exports = function checkUseres(message, args, num = 0) {
+module.exports = function checkRoles(message, args, num = 1) {
     if(args[num]) {
         if(!isNaN(parseInt(args[num])) && args[num].length >= 17){
-            if(message.guild.members.cache.get(args[num])) {
+            if(message.guild.roles.cache.get(args[num])) {
                 return args[num];
             } else return "not valid";
             
-        } else if(message.mentions.members.first()){
-            return message.mentions.members.first().id;
+        } else if(message.mentions.roles.first()){
+            return message.mentions.roles.first().id;
+		}else if (isNaN(parseInt(args[num]) && args[num].length >= 2 && args[num].length <= 32)) {
+			let Trole = message.guild.roles.cache.find(role => role.name === args.slice(1).join(' '));
+			if(Trole === undefined)return "not valid";
+			return Trole.id;
+
         }else if(message.mentions.everyone) {
             return "everyone";
-        }else if(args[num] === "me") {
-			return message.author.id;
-		} else return "not useable";
+        }else return "not useable";
     } else return "no args";
 }
 
+
+
 /*
-switch (checkUseres(message, args, 0)) {
+switch (checkRoles(message, args, 1)) {
 			case "not valid":
 			case "everyone":	
 			case "not useable":
@@ -44,4 +49,6 @@ switch (checkUseres(message, args, 0)) {
 				break;
 		}
 
- */
+
+
+*/
